Implemented bubble sort
Implemented the simplest sort algo that came to mind, bubble sort! Also got printing working as well. woohoo!

I want to implement other sorting algorithms, add a function that calculates statistics on the performance of each algorithm, and also come up with a strategy on my own.

I have yet to find an appropriate visualization method as well. Not only do I want to visualize the games, but having a visualization of performance of each algo as they each do
the tests could be cool.